+++ spec/features/all_projects/by_month/users_can_view_confrimed_projects_spec.rb
@@ -6,7 +6,7 @@ RSpec.feature "Users can view confirmed projects" do
     mock_all_academies_api_responses
     create(:conversion_project, conversion_date: Date.today.next_month.at_beginning_of_month, conversion_date_provisional: false)
 
-    visit confirmed_all_opening_projects_path
+    visit confirmed_all_by_month_projects_path
 
     within("thead") do
       expect(page).to have_content("Academy order type")