+++ spec/features/users_can_view_a_table_of_project_openers_spec.rb
@@ -13,7 +13,7 @@ RSpec.feature "Users can view project openers in table form" do
       it "shows the correct value" do
         _project = create(:conversion_project, conversion_date: Date.new(2023, 1, 1), conversion_date_provisional: false, all_conditions_met: true)
 
-        visit "/projects/all/opening/confirmed/1/2023"
+        visit "/projects/all/by-month/confirmed/1/2023"
         expect(page).to have_content("Yes")
       end
     end
@@ -22,7 +22,7 @@ RSpec.feature "Users can view project openers in table form" do
       it "shows the correct value" do
         _project = create(:conversion_project, conversion_date: Date.new(2023, 1, 1), conversion_date_provisional: false, all_conditions_met: nil)
 
-        visit "/projects/all/opening/confirmed/1/2023"
+        visit "/projects/all/by-month/confirmed/1/2023"
         expect(page).to have_content("Not yet")
       end
     end